+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 - fix caret of internal infoview at first line first col while refreshing
+- tuning size for internal infoview popup window
+- remove click in internal infoview, some optimization for popup up doc
+
 ## [0.0.17] - 2024-10-06
 
 Depend on the latest approved nightly build of LSP4IJ: 0.7.0-20241006-013203